requirements.txt
emoji
aiohttp
aiofiles
gTTS
dnspython
future
motor
pymongo[srv]
pillow
psutil
pykeyboard
Pyrogram==1.4.16
python-arq
requests
search_engine_parser
TgCrypto
uvloop
bs4
pyyaml==5.3.1
python-dotenv
feedparser
fuzzysearch
pytube
validators
heroku3
apscheduler
cachetools
regex
pytz
dateparser
gpytranslate==1.4.0
EmojiCaptcha
captcha
rich
gitpython
pytimeparse
googletrans==3.1.0a0
telegraph
